@@ -1,25 +1,25 @@
 name: flutterbook
 description: A storyboard tool to accelerate the development of Flutter widgets.
-version: 0.1.8
+version: 0.1.9
 homepage: https://github.com/GhostWalker562/flutterbook
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
+  sdk: ">=3.0.5 <4.0.0"
 
 dependencies:
   cupertino_icons: ^1.0.3
   device_preview: ^1.1.0
-  flutter_markdown: ^0.6.10+3
-  dart_code_viewer2: ^1.0.0
+  flutter_markdown: ^0.6.17+1
+  dart_code_viewer2: ^1.0.2
   flutter:
     sdk: flutter
   flutter_feather_icons: ^2.0.0+1
-  google_fonts: ^3.0.1
-  provider: "^6.0.1"
-  recase: ^4.0.0
+  google_fonts: ^6.1.0
+  provider: ^6.0.5
+  recase: ^4.1.0
 
 dev_dependencies:
-  flutter_lints: ^1.0.4
+  flutter_lints: ^3.0.1
   flutter_test:
     sdk: flutter
 
